fix: migrated error test, removed old resolve api
